+++ examples/cli/commands.py
@@ -63,7 +63,8 @@ def commands():
         "put-card-record":            Command(put_card_record,            [Args.controller, Args.card]),
         "delete-card":                Command(delete_card,                [Args.controller, Args.card]),
         "delete-all-cards":           Command(delete_all_cards,           [Args.controller]),
-        "get-event":                  Command(get_event,                  [Args.controller]),
+        "get-event":                  Command(get_event,                  [Args.controller, Args.index]),
+        "get-event-record":           Command(get_event_record,           [Args.controller, Args.index]),
         "get-event-index":            Command(get_event_index,            [Args.controller]),
         "set-event-index":            Command(set_event_index,            [Args.controller]),
         "record-special-events":      Command(record_special_events,      [Args.controller]),
@@ -374,7 +375,7 @@ def get_event(u, dest, timeout, args, protocol="udp"):
     Retrieves the information for an event from a controller using the 'get_event' API function.
     """
     controller = (args.controller, dest, protocol)
-    index = EVENT_INDEX
+    index = args.index
 
     response = u.get_event(controller, index, timeout=timeout)
 
@@ -385,6 +386,16 @@ def get_event(u, dest, timeout, args, protocol="udp"):
         raise ValueError(f"event @ index {index} not found")
 
     return response
+
+
+def get_event_record(u, dest, timeout, args, protocol="udp"):
+    """
+    Retrieves the information for an event from a controller using the 'get_event' API function.
+    """
+    controller = (args.controller, dest, protocol)
+    index = args.index
+
+    return u.get_event(controller, index, timeout=timeout)
 
 
 def get_event_index(u, dest, timeout, args, protocol="udp"):