+++ integration-tests/uhppoted/udp_default.py
@@ -19,6 +19,8 @@ from ipaddress import IPv4Address
 from uhppoted import uhppote
 from uhppoted.net import dump
 from uhppoted.structs import Card
+from uhppoted.errors import EventNotFound
+from uhppoted.errors import EventOverwritten
 
 # pylint: disable=relative-beyond-top-level
 from .stub import messages
@@ -28,6 +30,8 @@ CONTROLLER = 405419896
 CARD = 8165538
 CARD_INDEX = 2
 EVENT_INDEX = 29
+EVENT_INDEX_NOT_FOUND = 200
+EVENT_INDEX_OVERWRITTEN = 73
 TIME_PROFILE = 29
 
 
@@ -350,6 +354,35 @@ class TestUDPWithDestAddr(unittest.TestCase):
         response = self.u.get_event(controller, index)
 
         self.assertEqual(response, expected.GetEventResponse)
+
+    def test_get_event_record(self):
+        """
+        Tests the get-event-record function with defaults.
+        """
+        controller = CONTROLLER
+        index = EVENT_INDEX
+
+        record = self.u.get_event_record(controller, index)
+
+        self.assertEqual(record, expected.GetEventRecord)
+
+    def test_get_event_record_not_found(self):
+        """
+        Tests the get-event-record function for a non-existent record.
+        """
+        controller = CONTROLLER
+        index = EVENT_INDEX_NOT_FOUND
+
+        self.assertRaises(EventNotFound, self.u.get_event_record, controller, index)
+
+    def test_get_event_record_overwritten(self):
+        """
+        Tests the get-event-record function for a record that has been overwritten.
+        """
+        controller = 201020304
+        index = EVENT_INDEX_OVERWRITTEN
+
+        self.assertRaises(EventOverwritten, self.u.get_event_record, controller, index)
 
     def test_get_event_index(self):
         """